# backend/internal/modules/auth/infrastructure/http/handler.go
package http

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/lukcba/club-pulse-system-api/backend/internal/core/errors"
	"github.com/lukcba/club-pulse-system-api/backend/internal/modules/auth/application"
)

type AuthHandler struct {
	useCase *application.AuthUseCases
}

func NewAuthHandler(useCase *application.AuthUseCases) *AuthHandler {
	return &AuthHandler{
		useCase: useCase,
	}
}

func (h *AuthHandler) Register(c *gin.Context) {
	var dto application.RegisterDTO
	if err := c.ShouldBindJSON(&dto); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input"})
		return
	}

	token, err := h.useCase.Register(dto)
	if err != nil {
		handleError(c, err)
		return
	}

	c.JSON(http.StatusCreated, token)
}

// Login godoc
// @Summary      Login user
// @Description  Authenticate user and return HttpOnly cookies
// @Tags         auth
// @Accept       json
// @Produce      json
// @Param        input body application.LoginDTO true "Login Credentials"
// @Success      200   {object}  map[string]string "message: Login successful"
// @Failure      400   {object}  map[string]string
// @Router       /auth/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var dto application.LoginDTO
	if err := c.ShouldBindJSON(&dto); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Invalid input"})
		return
	}

	token, err := h.useCase.Login(dto)
	if err != nil {
		handleError(c, err)
		return
	}

	// Set HttpOnly Cookie
	// MaxAge: 24 hours (86400 seconds)
	// Path: "/"
	// Domain: "" (localhost handles this)
	// Secure: false (for dev, true in prod)
	// HttpOnly: true
	c.SetCookie("access_token", token.AccessToken, 86400, "/", "", false, true)

	// We can also set refresh token if available
	if token.RefreshToken != "" {
		c.SetCookie("refresh_token", token.RefreshToken, 86400*7, "/", "", false, true)
	}

	c.JSON(http.StatusOK, gin.H{"message": "Login successful"})
}

func (h *AuthHandler) RefreshToken(c *gin.Context) {
	type RefreshDTO struct {
		RefreshToken string `json:"refresh_token" binding:"required"`
	}
	var dto RefreshDTO
	if err := c.ShouldBindJSON(&dto); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Refresh token required"})
		return
	}

	token, err := h.useCase.RefreshToken(dto.RefreshToken)
	if err != nil {
		handleError(c, err)
		return
	}

	c.JSON(http.StatusOK, token)
}

func (h *AuthHandler) Logout(c *gin.Context) {
	type LogoutDTO struct {
		RefreshToken string `json:"refresh_token" binding:"required"`
	}
	var dto LogoutDTO
	if err := c.ShouldBindJSON(&dto); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Refresh token required"})
		return
	}

	if err := h.useCase.Logout(dto.RefreshToken); err != nil {
		handleError(c, err)
		return
	}

	c.Status(http.StatusNoContent)
}

func (h *AuthHandler) ListSessions(c *gin.Context) {
	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}

	sessions, err := h.useCase.ListUserSessions(userID.(string))
	if err != nil {
		handleError(c, err)
		return
	}

	c.JSON(http.StatusOK, sessions)
}

func (h *AuthHandler) RevokeSession(c *gin.Context) {
	sessionID := c.Param("id")
	if sessionID == "" {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Session ID required"})
		return
	}

	userID, exists := c.Get("userID")
	if !exists {
		c.JSON(http.StatusUnauthorized, gin.H{"error": "Unauthorized"})
		return
	}

	if err := h.useCase.RevokeSession(sessionID, userID.(string)); err != nil {
		handleError(c, err)
		return
	}

	c.JSON(http.StatusOK, gin.H{"message": "Session revoked"})
}

func (h *AuthHandler) GoogleLogin(c *gin.Context) {
	type GoogleLoginDTO struct {
		Code string `json:"code" binding:"required"`
	}
	var dto GoogleLoginDTO
	if err := c.ShouldBindJSON(&dto); err != nil {
		c.JSON(http.StatusBadRequest, gin.H{"error": "Code is required"})
		return
	}

	token, err := h.useCase.GoogleLogin(c.Request.Context(), dto.Code)
	if err != nil {
		handleError(c, err)
		return
	}

	// Set HttpOnly Cookie
	// MaxAge: 24 hours (86400 seconds)
	c.SetCookie("access_token", token.AccessToken, 86400, "/", "", false, true)
	if token.RefreshToken != "" {
		c.SetCookie("refresh_token", token.RefreshToken, 86400*7, "/", "", false, true)
	}

	c.JSON(http.StatusOK, gin.H{"message": "Google login successful"})
}

// Routes registration
// Routes registration
// Routes registration
func RegisterRoutes(r *gin.RouterGroup, h *AuthHandler, authMiddleware gin.HandlerFunc) {
	authGroup := r.Group("/auth")
	{
		authGroup.POST("/register", h.Register)
		authGroup.POST("/login", h.Login)
		authGroup.POST("/refresh", h.RefreshToken)
		authGroup.POST("/logout", h.Logout)
		authGroup.POST("/google", h.GoogleLogin)
	}

	// Protected routes
	protected := authGroup.Group("")
	protected.Use(authMiddleware)
	{
		protected.GET("/sessions", h.ListSessions)
		protected.DELETE("/sessions/:id", h.RevokeSession)
	}
}

// Simple error handler helper (move to core/platform later if needed shared)
func handleError(c *gin.Context, err error) {
	if e, ok := err.(*errors.AppError); ok {
		c.JSON(e.Code, gin.H{"error": e.Message, "type": e.Type})
		return
	}
	c.JSON(http.StatusInternalServerError, gin.H{"error": "Internal server error"})
}